serv: Добавлено 2 API контроллера для корзины  и вишлиста, убран неиспользуемый код во всем проекте
