+++ server/BLL/ClothDTOs/AdminDTOs/AdminDTO.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace ClothDTOs;
 
 public class AdminDTO
