+++ server/BLL/ClothDTOs/AdminDTOs/AdminUpdateDTO.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace ClothDTOs;
 
 public class AdminUpdateDTO
