+++ server/BLL/ClothDTOs/CartDTOs/CartItemDTO.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Text.Json.Serialization;
 
 namespace ClothDTOs;
