+++ server/BLL/ClothDTOs/WishlistDTOs/WishlistItemDTO.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Text.Json.Serialization;
 
 namespace ClothDTOs.WishlistDTOs;
