+++ server/BLL/ClothesInterfacesBLL/IAuthAdminService.cs
@@ -1,4 +1,3 @@
-using System;
 using ClothDTOs;
 
 namespace ClothesInterfacesBLL;
