+++ server/BLL/ClothesInterfacesBLL/IAuthService.cs
@@ -1,5 +1,3 @@
-using System;
-using ClothDomain;
 using ClothDTOs;
 
 namespace ClothesInterfacesBLL;
