+++ server/BLL/ClothesInterfacesBLL/ICartBLL.cs
@@ -1,4 +1,3 @@
-using System;
 using ClothDTOs;
 
 namespace ClothesInterfacesBLL;
@@ -6,7 +5,6 @@ namespace ClothesInterfacesBLL;
 public interface ICartBLL
 {
     public Task<List<CartItemDTO>> GetCartItems();
-    public Task AddToCart(Guid clothId);
     public Task AddAmountOfCartItem(Guid cartId);
     public Task ReduceAmountOfCartItem(Guid cartId);
     public Task DeleteCartItem(Guid cartId);
