+++ server/BLL/ClothesInterfacesBLL/IWishlistBLL.cs
@@ -1,4 +1,3 @@
-using System;
 using ClothDTOs.WishlistDTOs;
 
 namespace ClothesInterfacesBLL;
@@ -6,7 +5,6 @@ namespace ClothesInterfacesBLL;
 public interface IWishlistBLL
 {
     public Task<List<WishlistItemDTO>> GetWishlistItems();
-    public Task AddToWishlist(Guid clothId);
     public Task FromWishlistToCart(Guid wishlistId);
     public Task DeleteWishlistItem(Guid wishlistId);
 }
