+++ server/Core/ClothDomain/Wishlist.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace ClothDomain;
 
 public class Wishlist
