+++ server/DAL/ClothesInterfacesDAL/ICartDAO.cs
@@ -1,4 +1,3 @@
-using System;
 using ClothDomain;
 
 namespace ClothesInterfacesDAL;
@@ -6,7 +5,6 @@ namespace ClothesInterfacesDAL;
 public interface ICartDAO
 {
     public Task<List<CartItem>> GetCartItems(Guid buyerId);
-    public Task AddToCart(Guid buyerId, Guid clothId);
     public Task AddAmountOfCartItem(Guid buyerId, Guid cartId);
     public Task ReduceAmountOfCartItem(Guid buyerId, Guid cartId);
     public Task DeleteCartItem(Guid buyerId, Guid cartId);
