+++ server/DAL/ClothesInterfacesDAL/IWishlistDAO.cs
@@ -1,4 +1,3 @@
-using System;
 using ClothDomain;
 
 namespace ClothesInterfacesDAL;
@@ -6,7 +5,6 @@ namespace ClothesInterfacesDAL;
 public interface IWishlistDAO
 {
     public Task<List<WishlistItem>> GetWishlistItems(Guid buyerId);
-    public Task AddToWishlist(Guid buyerId, Guid clothId);
     public Task FromWishlistToCart(Guid buyerId, Guid wishlistId);
     public Task DeleteWishlistItem(Guid buyerId, Guid wishlistId);
 }
