+++ server/DAL/ClothingStorePersistence/SqlBuyersDAO.cs
@@ -1,4 +1,3 @@
-using System;
 using ClothDomain;
 using ClothesInterfacesDAL;
 using Microsoft.EntityFrameworkCore;
