+++ server/DAL/ClothingStorePersistence/SqlDAO.cs
@@ -1,11 +1,10 @@
-using System;
 using ClothDomain;
 using ClothesInterfacesDAL;
 using Microsoft.EntityFrameworkCore;
 
 namespace ClothingStorePersistence;
 
-public class SqlDAO : IClothesDAO, ICartDAO, IWishlistDAO
+public class SqlDAO : IClothesDAO
 {
     private readonly ApplicationDbContext _context;
 
@@ -81,70 +80,6 @@ public class SqlDAO : IClothesDAO, ICartDAO, IWishlistDAO
         await _context.SaveChangesAsync();
     }
 
-    public async Task<List<CartItem>> GetCartItems(Guid buyerId){
-        var cart = await _context.Carts.Include(c => c.Items).ThenInclude(ci => ci.Cloth).FirstOrDefaultAsync(c => c.BuyerId == buyerId);
-        if(cart == null){
-            throw new Exception("Корзина не найдена.");
-        }
-
-        return cart.Items.ToList();
-    }
-
-    public async Task AddAmountOfCartItem(Guid buyerId, Guid cartId){
-        var cart = await _context.Carts.Include(c => c.Items).FirstOrDefaultAsync(c => c.BuyerId == buyerId);
-        if(cart == null){
-            throw new Exception("Корзина не найдена.");
-        }
-
-        var existingItem = cart.Items.FirstOrDefault(ci => ci.CartId == cartId);
-        if(existingItem == null){
-            throw new Exception("Продукт не найден.");
-        }
-        else{
-            existingItem.Amount++;
-        }
-
-        await _context.SaveChangesAsync();
-    }
-
-    public async Task ReduceAmountOfCartItem(Guid buyerId, Guid cartId){
-        var cart = await _context.Carts.Include(c => c.Items).FirstOrDefaultAsync(c => c.BuyerId == buyerId);
-        if(cart == null){
-            throw new Exception("Корзина не найдена.");
-        }
-
-        var existingItem = cart.Items.FirstOrDefault(ci => ci.CartId == cartId);
-        if(existingItem == null){
-            throw new Exception("Продукт не найден.");
-        }
-        else{
-            existingItem.Amount--;
-        }
-
-        if(existingItem.Amount == 0){
-            _context.CartItems.Remove(existingItem);
-        }
-
-        await _context.SaveChangesAsync();
-    }
-
-    public async Task DeleteCartItem(Guid buyerId, Guid cartId){
-        var cart = await _context.Carts.Include(c => c.Items).FirstOrDefaultAsync(c => c.BuyerId == buyerId);
-        if(cart == null){
-            throw new Exception("Корзина не найдена");
-        }
-
-        var existingItem = cart.Items.FirstOrDefault(ci => ci.CartId == cartId);
-        if(existingItem == null){
-            throw new Exception("Продукт не найден.");
-        }
-        else{
-            _context.CartItems.Remove(existingItem);
-        }
-
-        await _context.SaveChangesAsync();
-    }
-
     public async Task AddToWishlist(Guid buyerId, Guid clothId){
         var wishlist = await _context.Wishlists.Include(c => c.Items).FirstOrDefaultAsync(c => c.BuyerId == buyerId);
         if(wishlist == null){
@@ -160,44 +95,4 @@ public class SqlDAO : IClothesDAO, ICartDAO, IWishlistDAO
         wishlist.Items.Add(new WishlistItem { ClothId = clothId });
         await _context.SaveChangesAsync();
     }
-
-    public async Task<List<WishlistItem>> GetWishlistItems(Guid buyerId){
-        var wishlist = await _context.Wishlists.Include(c => c.Items).ThenInclude(ci => ci.Cloth).FirstOrDefaultAsync(c => c.BuyerId == buyerId);
-        if(wishlist == null){
-            throw new Exception("Вишлист не найден.");
-        }
-
-        return wishlist.Items.ToList();
-    }
-
-    public async Task FromWishlistToCart(Guid buyerId, Guid wishlistId){
-        var wishlist = await _context.Wishlists.Include(c => c.Items).FirstOrDefaultAsync(c => c.BuyerId == buyerId);
-        if(wishlist == null){
-            throw new Exception("Вишлист не найден.");
-        }
-
-        var existingWishlistItem = wishlist.Items.FirstOrDefault(ci => ci.WishlistId == wishlistId);
-        if(existingWishlistItem == null){
-            throw new Exception("Продукт не найден.");
-        }
-        
-        await AddToCart(buyerId, existingWishlistItem.ClothId);
-        _context.WishlistItems.Remove(existingWishlistItem);
-        await _context.SaveChangesAsync();
-    }
-
-    public async Task DeleteWishlistItem(Guid buyerId, Guid wishlistId){
-        var wishlist = await _context.Wishlists.Include(c => c.Items).FirstOrDefaultAsync(c => c.BuyerId == buyerId);
-        if(wishlist == null){
-            throw new Exception("Вишлист не найден");
-        }
-
-        var existingItem = wishlist.Items.FirstOrDefault(ci => ci.WishlistId == wishlistId);
-        if(existingItem == null){
-            throw new Exception("Продукт не найден.");
-        }
-
-        _context.WishlistItems.Remove(existingItem);
-        await _context.SaveChangesAsync();
-    } 
 }
