+++ server/PL/clothingStoreWebAPI/Controllers/AuthControllers/AuthAdminController.cs
@@ -1,4 +1,3 @@
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using ClothesInterfacesBLL;
 using ClothDTOs;
