+++ server/PL/clothingStoreWebAPI/Controllers/AuthControllers/AuthController.cs
@@ -1,7 +1,5 @@
 using ClothDTOs;
 using ClothesInterfacesBLL;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace clothingStoreWebAPI.Controllers
