+++ server/PL/clothingStoreWebAPI/Controllers/BuyerController.cs
@@ -1,7 +1,6 @@
 using ClothDTOs;
 using ClothesInterfacesBLL;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace clothingStoreWebAPI.Controllers
