+++ server/PL/clothingStoreWebAPI/Models/LoginModel.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace clothingStoreWebAPI.Controllers;
 
 public class LoginModel
